backend/requirements.txt
Flask==3.0.3
Flask-Cors==4.0.1
Flask-SQLAlchemy==3.1.1
PyJWT==2.9.0
cryptography==43.0.3
requests==2.32.3
python-dotenv==1.0.1
google-auth==2.35.0

# --- Optional production WSGI server ------------------------------------
# Uncomment for `gunicorn -w 4 -b 0.0.0.0:5001 'app:app'` deployments.
# In dev `python app.py` does not require it.
# gunicorn==23.0.0
